Changed StationId from int to long

+++ EveLib.EveXml/Models/Misc/ConquerableStations.cs
@@ -40,7 +40,7 @@ namespace eZet.EveLib.EveXmlModule.Models.Misc {
             /// </summary>
             /// <value>The station identifier.</value>
             [XmlAttribute("stationID")]
-            public int StationId { get; set; }
+            public long StationId { get; set; }
 
             /// <summary>
             ///     Gets or sets the name of the station.
